finger_counting_model/requirements.txt
tensorflow==2.14.0
numpy==1.26.4
scikit-learn==1.3.0
matplotlib==3.7.2
opencv-python==4.8.1.78
pillow==9.4.0
joblib==1.2.0